moving hwi_rpc.py
